Move `RocksDbBuilder` into separate module
