+++ lib/segment/src/segment_constructor/mod.rs
@@ -1,3 +1,4 @@
+mod rocksdb_builder;
 pub mod segment_builder;
 mod segment_constructor_base;
 pub mod simple_segment_constructor;
